move plugin download its own file
